Cleaning up repo for merging RTCC code.
